+++ .storybook/config.js
@@ -35,6 +35,7 @@ configure(() => {
   require('./../packages/bpk-component-code/stories');
   require('./../packages/bpk-component-content-container/stories');
   require('./../packages/bpk-component-datepicker/stories');
+  require('./../packages/bpk-component-datatable/stories');
   require('./../packages/bpk-component-fieldset/stories');
   require('./../packages/bpk-component-form-validation/stories');
   require('./../packages/bpk-component-grid/stories');